+++ veecle-freertos-integration/tests/queue_async_messages_waiting.rs
@@ -1,7 +1,7 @@
 #![expect(missing_docs)]
 
-use freertos_rust::{CurrentTask, Duration, Task, TaskPriority, channel};
 use futures::FutureExt;
+use veecle_freertos_integration::{CurrentTask, Duration, Task, TaskPriority, channel};
 
 pub mod common;