+++ veecle-freertos-integration/tests/queue_async_send_from_isr_exceed_max_capacity.rs
@@ -1,6 +1,6 @@
 #![expect(missing_docs)]
 
-use freertos_rust::{InterruptContext, channel};
+use veecle_freertos_integration::{InterruptContext, channel};
 
 pub mod common;