+++ veecle-freertos-integration/tests/queue_receive_no_send.rs
@@ -1,6 +1,6 @@
 #![expect(missing_docs)]
 
-use freertos_rust::{Duration, FreeRtosError, Queue};
+use veecle_freertos_integration::{Duration, FreeRtosError, Queue};
 
 pub mod common;