+++ veecle-freertos-integration/tests/queue_send_exceed_max_capacity.rs
@@ -1,6 +1,6 @@
 #![expect(missing_docs)]
 
-use freertos_rust::{Duration, Queue};
+use veecle_freertos_integration::{Duration, Queue};
 
 pub mod common;