+++ veecle-freertos-integration/tests/task_wait_for_notification.rs
@@ -2,7 +2,7 @@
 
 use std::sync::atomic::{AtomicBool, Ordering};
 
-use freertos_rust::{CurrentTask, Duration, TaskNotification};
+use veecle_freertos_integration::{CurrentTask, Duration, TaskNotification};
 
 use crate::common::start_task;
 
@@ -35,7 +35,7 @@ fn task_wait_for_notification() {
         CurrentTask::suspend();
     });
 
-    freertos_rust::scheduler::start_scheduler();
+    veecle_freertos_integration::scheduler::start_scheduler();
 
     assert!(WAITED.load(Ordering::Acquire));
 }