+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Next
 
+## veecle-freertos-integration
+
+* Remove explicit library naming.
+
 # 0.1.0
 
 * Initial release.